Added a wait time functionality
Wait time functionality works in two options. After stopping the mouse and keyboard event listeners, if the wait time is already known, an input bar has been put to wait that long. If not, users can use the "start waiting" and "stop waiting" buttons to measure and record the amount of time to wait.